Fix warning about require with an expression instead of a string
`require` call for 'xmlhttprequest' in `makeRequest.js` is called with a variable instead of a string, causing warnings in build tools.

```
warning  in ./~/iota.lib.js/lib/utils/makeRequest.js
10:16-46 Critical dependency: the request of a dependency is an expression
```
Simply removing the variable and passing the string directly fixes that. There is no point in caching that string in a variable anyway.